Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.2.1
torchvision==0.17.1
timm==0.9.2
pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
google-cloud-vision==3.4.4
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
python-magic==0.4.27
google-api-python-client==2.107.0
serpapi==0.1.0 
beautifulsoup4==4.12.3
yt-dlp